refactor: separate activation and config responsibilities
