+++ crates/protocol-diff/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "vpnhide_protocol_diff"
+version = "0.0.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+publish = false
+
+[dependencies]
+vpnhide_protocol = { path = "../protocol" }
+
+[dev-dependencies]
+proptest = "1.6"
+
+[build-dependencies]
+cc = "1.2"
