+++ crates/protocol-diff/c/config_wrapper.c
@@ -1,0 +1,10 @@
+#include "shared/vpnhide_logic.h"
+
+/* Stable external symbol for the otherwise static-inline freestanding parser. */
+int vpnhide_diff_parse_config(const unsigned char *input, unsigned long len,
+                              struct vpnhide_target *targets, int capacity,
+                              int *debug)
+{
+    return vpnhide_parse_config((const char *)input, len, targets, capacity,
+                                debug);
+}
